pequenas modificações no service e exceptions do livro, teste do livroService feito

+++ src/main/java/com/Weimar_Jr/SistemaDeSugestoesEFeedbackDeLivrosDaBibliotecaEscolar/Entidades/Biblioteca/Feedback.java
@@ -22,7 +22,7 @@ public class Feedback {
     @NotNull
     private String comentario;
     @NotNull
-    private int avaliacao;
+    private Double avaliacao;
     @NotNull
     private Boolean nomeDoAlunoVisivel;
